update an ke dua

+++ php/favorite_toggle.php
@@ -7,21 +7,31 @@ if (!isset($_SESSION["nm_user"])) {
     exit();
 }
 
-$username = $_SESSION["nm_user"];
-$idResep = $_POST['id_resep'];
+if (isset($_GET['halFav'])) {
+    $username = $_SESSION["nm_user"];
+    $idResep = $_POST['id_resep'];
+    $halFav = $_GET['halFav'];
 
-// Cek apakah sudah ada
-$cek = $koneksi->query("SELECT * FROM favorit WHERE username = '$username' AND id_resep = $idResep");
+    // Cek apakah sudah ada
+    $cek = $koneksi->query("SELECT * FROM favorit WHERE username = '$username' AND id_resep = $idResep");
 
-if ($cek->num_rows > 0) {
-    // Kalau sudah ada → hapus dari favorit
-    $koneksi->query("DELETE FROM favorit WHERE username = '$username' AND id_resep = $idResep");
-    echo "<script>window.location='../halaman_resep.php?id=$idResep';</script>";
-} else {
-    // Kalau belum ada → tambahkan
-    $koneksi->query("INSERT INTO favorit (username, id_resep, created_at) VALUES ('$username', $idResep, NOW())");
-    echo "<script>window.location='../halaman_resep.php?id=$idResep';</script>";
+    if ($cek->num_rows > 0) {
+        // Kalau sudah ada → hapus dari favorit
+        $koneksi->query("DELETE FROM favorit WHERE username = '$username' AND id_resep = $idResep");
+        if ($halFav === 'no') {
+            echo "<script>window.location='../halaman_resep.php?id=$idResep';</script>";
+        } else {
+            echo "<script>window.location='../halaman_resep_fav.php?id=$idResep';</script>";
+        }
+    } else {
+        // Kalau belum ada → tambahkan
+        $koneksi->query("INSERT INTO favorit (username, id_resep, created_at) VALUES ('$username', $idResep, NOW())");
+        if ($halFav === 'no') {
+            echo "<script>window.location='../halaman_resep.php?id=$idResep';</script>";
+        } else {
+            echo "<script>window.location='../halaman_resep_fav.php?id=$idResep';</script>";
+        }
+    }
+
+    exit();
 }
-
-// header("Location: ../halaman_detail_resep_admin.php?id=$idResep");
-exit();
